+++ turbo/apps/api/src/signals/routes/zero-runs-cancel.ts
@@ -12,7 +12,7 @@ import {
   dispatchCancelSideEffects$,
   shouldDispatchCancelSideEffects,
   type CancelRunResult,
-} from "../services/zero-run-cancel.service";
+} from "../services/run-cancel.service";
 import { tapError } from "../utils";
 import type { RouteEntry } from "../route-entry";
