+++ turbo/apps/api/src/signals/services/agent-run-create.service.ts
@@ -273,7 +273,7 @@ import {
   checkOrgCreditsForRunAdmission,
   checkResolvedOrgCreditsForRunAdmission,
   resolveOrgCreditAvailability,
-} from "./zero-run-admission.service";
+} from "./run-admission.service";
 import { activateUsageAllowanceWindowsForRun } from "./usage-allowance.service";
 import {
   ApiDispatchPhaseCollector,
