+++ turbo/apps/api/src/signals/services/agent-webhook-complete.service.ts
@@ -37,7 +37,7 @@ import { lockChatQueueThread } from "./chat-event-queue.service";
 import { projectLegacyCheckpointStorage } from "./storage-legacy-projection.service";
 import { maybeEmitRunUsageEvent$ } from "./chat-usage-event.service";
 import { processOrgUsageEvents$ } from "./credit-usage.service";
-import { drainOrgQueue$ } from "./zero-run-queue.service";
+import { drainOrgQueue$ } from "./run-queue.service";
 
 type WebhookCompleteBody = z.infer<
   typeof webhookCompleteContract.complete.body
