+++ turbo/apps/api/src/signals/services/threadless-run-cleanup.service.ts
@@ -25,11 +25,8 @@ import { writeDb$, type Db } from "../external/db";
 import { settle } from "../utils";
 import { failPendingInlineOnlyDeliveryCallbacksForDeletedThread } from "./agent-run-callback.service";
 import { dispatchCompleteSideEffects$ } from "./agent-webhook-complete.service";
-import {
-  cancelRun$,
-  dispatchCancelSideEffects$,
-} from "./zero-run-cancel.service";
-import { drainOrgQueue$ } from "./zero-run-queue.service";
+import { cancelRun$, dispatchCancelSideEffects$ } from "./run-cancel.service";
+import { drainOrgQueue$ } from "./run-queue.service";
 
 const L = logger("ThreadlessRunCleanup");
