+++ turbo/apps/api/src/signals/services/zero-chat-thread.service.ts
@@ -57,7 +57,7 @@ import {
   appendChatThreadEvent,
   chatThreadServiceTierFromCodex,
 } from "./zero-chat-thread-event.service";
-import { cancelRun$, type CancelRunResult } from "./zero-run-cancel.service";
+import { cancelRun$, type CancelRunResult } from "./run-cancel.service";
 import { runOwnedChatEventForRunCondition } from "./chat-event-type.service";
 import { cancellationRecoveryPendingForThread } from "./chat-active-run.service";
