+++ turbo/apps/api/src/signals/services/zero-model-selection.service.ts
@@ -30,7 +30,7 @@ import { and, eq, or } from "drizzle-orm";
 import { badRequestMessage, insufficientCredits } from "../../lib/error";
 import type { Db } from "../external/db";
 import { ensureOrgModelPolicies } from "./zero-model-policy.service";
-import { checkOrgCreditsForRunAdmission } from "./zero-run-admission.service";
+import { checkOrgCreditsForRunAdmission } from "./run-admission.service";
 import {
   loadOrgPlanCapabilities,
   type OrgPlanCapabilities,
